Reduce chance of inadvertent modification of current version detection
Although this does not completely remove the possibility of other software
interfering with the variable used to identify the latest version installed
this does at least use a more unique variable to store the version
identified by the installer as being the current version.

Created a new variable $sitemap_current_version that is then used elsewhere
outside of the immediate installer as it is outside of this installer that
the variable may be re-assigned.  An alternative to this would be to again
call the installer software as part of loading the page so that the
variable would in effect be defined within the current code not separated
by other executable code.