Add text for TEXT_SITEMAPXML_TIPS_TEXT when not defined.
If the database uninstall is performed (or before this has been installed),
the TEXT_SITEMAPXML_TIPS_TEXT is not defined because it depends on data
in the database that has not yet been set.  In an installation evolution
though the data has likely just been put into the database or is about to
be on the next load cycle so this offers some information to accompany that
condition.  Really what it does is prevent a future error by ensuring that
the content is defined.  I had considered incorporating another define, but
this seems more appropriate to offer alternate text to the template instead
of the template being clouded with various logic.